+++ test/test_test.dart
@@ -1,9 +1,0 @@
-import 'package:test/test.dart';
-
-void main() {
-  group('A test', () {
-    test('Passing test', () {
-      expect(1, equals(1));
-    });
-  });
-}